Argument 'check' can be issued to just validate the documentation.
svn path=/trunk/l10n-support/pology/; revision=1163628

+++ doc/user/makedoc.sh
@@ -2,9 +2,13 @@
 
 cd $(dirname $0)
 
+mode=$1
+
 topdbk=top.docbook
 
 xmllint --noout --xinclude --postvalid $topdbk || exit 1
+
+test "$mode" = "check" && exit 0
 
 rm -rf html/*; mkdir -p html
 ln -s ../html-data html/data
